Move unit test TestExpmInCircuit to new module
Creates a new unit-test module `test_jax_linalg_in_circuit.py``
specifically for testing JAX linear algebra functions in full quantum-
circuit workflows compiled with qjit.
